I added the ability to change the prefix without affecting the "global" one, so if I want it to be like "[Console] " normally I could also create a prefix like :s for "<Strahan> " so I could make it look like I was chatting without having to keep changing the primary prefix.
I also changed some permission nodes to make them more logical and some
message keys so if you are upgrading, be sure you verify the perms.
IIRC, it was just the toggle that changed I think. Since there are more
toggles now, I made the :toggle command now :console to toggle console
and the perm became hierarchical so consolechat.toggle.console and
seeglobal.

Which brings me to the next change, :seeglobal. When you send a global
msg, it didn't show you the output. I figured maybe ppl would like to
see it so now you can toggle whether or not you do. It's not player
specific, it's global. If someone really wants it to be player specific,
lemme know and I'll add that.